Split shared cache in backend and frontend
