+++ executorlib/backend/cache_serial.py
@@ -1,6 +1,6 @@
 import sys
 
-from executorlib.shared.cache import execute_task_in_file
+from executorlib.cache.backend import backend_execute_task_in_file
 
 if __name__ == "__main__":
-    execute_task_in_file(file_name=sys.argv[1])
+    backend_execute_task_in_file(file_name=sys.argv[1])
